Fix controls gptokeyb
gptokeyb is creating a virtual input device sometimes after the port launch itself.
This is a problem specific to TSP's stock firmware: a new input device can't be detected after the launch of an app.
The result of this behavior is that many ports doesn't have their controls working (quake 3, fallout,
To fix that we add a short sleep time in the port's launch script just after the gptokeyb command line.

---------

We are parsing "/tmp/log/messages" from stock firmware to determine which option has been selected in the emulator menu selector. This allows to maintain only one launch script.